requirements.txt
# -----------------------------
# API Framework
# -----------------------------
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.3

# -----------------------------
# Database (SYNC)
# -----------------------------
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1

# -----------------------------
# Queue System
# -----------------------------
redis==5.0.1
rq==1.16.1

# -----------------------------
# Scraping
# -----------------------------
playwright==1.42.0
beautifulsoup4==4.12.3
lxml==5.1.0
httpx==0.27.0

# -----------------------------
# Config
# -----------------------------
python-dotenv==1.0.1
pydantic-settings==2.2.1

# -----------------------------
# Utilities
# -----------------------------
fake-useragent==1.5.1
tenacity==8.2.3

# -----------------------------
# Logging
# -----------------------------
structlog==24.1.0

# -----------------------------
# Scheduling
# -----------------------------
APScheduler==3.10.4

# -----------------------------
# Testing
# -----------------------------
pytest==8.1.1
pytest-asyncio==0.23.5

# -----------------------------
# Dev
# -----------------------------
black==24.3.0
isort==5.13.2